moved to functional instead of unit test
