+++ pytest.ini
@@ -1,9 +1,10 @@
 [pytest]
 markers =
     acceptance: marks tests as acceptance tests
+    functional: marks tests as functional tests
     unit:       marks tests as unit tests
 
-testpaths = tests/acceptance tests/unit
+testpaths = tests/acceptance tests/functional tests/unit
 pythonpath = functions layers/utils
 
 filterwarnings =
